+++ cpp/src/aztec3/circuits/rollup/base/utils.cpp
@@ -1,3 +1,4 @@
+#include "aztec3/circuits/rollup/base/nullifier_tree_testing_harness.hpp"
 #include "index.hpp"
 #include "init.hpp"
 
@@ -97,6 +98,128 @@ PreviousRollupData<NT> dummy_previous_rollup_with_vk_proof()
     };
 
     return previous_rollup;
+}
+
+NullifierMemoryTreeTestingHarness get_initial_nullifier_tree(size_t spacing = 5)
+{
+    // Create a nullifier tree with 8 nullifiers, this padding is required so that the default 0 value in an indexed
+    // merkle tree does not affect our tests Nullifier tree at the start
+    NullifierMemoryTreeTestingHarness nullifier_tree = NullifierMemoryTreeTestingHarness(NULLIFIER_TREE_HEIGHT);
+    // Start from 1 as 0 is always inserted
+    for (size_t i = 1; i < 8; ++i) {
+        // insert 5, 10, 15, 20 ...
+        nullifier_tree.update_element(i * spacing);
+    }
+    return nullifier_tree;
+}
+
+AppendOnlyTreeSnapshot<NT> get_snapshot_of_tree_state(NullifierMemoryTreeTestingHarness nullifier_tree)
+{
+    return {
+        .root = nullifier_tree.root(),
+        .next_available_leaf_index = uint32_t(8),
+    };
+}
+
+std::tuple<BaseRollupInputs<NT>, AppendOnlyTreeSnapshot<NT>, AppendOnlyTreeSnapshot<NT>>
+generate_nullifier_tree_testing_values(BaseRollupInputs<NT> inputs,
+                                       size_t starting_insertion_index = 0,
+                                       size_t spacing = 5)
+{
+
+    // Generate nullifier tree testing values
+
+    NullifierMemoryTreeTestingHarness nullifier_tree = get_initial_nullifier_tree(spacing);
+    NullifierMemoryTreeTestingHarness parallel_insertion_tree = get_initial_nullifier_tree(spacing);
+
+    AppendOnlyTreeSnapshot<NT> nullifier_tree_start_snapshot = get_snapshot_of_tree_state(nullifier_tree);
+
+    const size_t NUMBER_OF_NULLIFIERS = KERNEL_NEW_NULLIFIERS_LENGTH * 2;
+    std::array<NullifierLeafPreimage<NT>, NUMBER_OF_NULLIFIERS> new_nullifier_leaves;
+    std::array<NullifierLeafPreimage<NT>, NUMBER_OF_NULLIFIERS> low_nullifier_leaves_preimages;
+    std::array<MembershipWitness<NT, NULLIFIER_TREE_HEIGHT>, NUMBER_OF_NULLIFIERS>
+        low_nullifier_leaves_preimages_witnesses;
+
+    // Calculate the low nullifier pre-images (spoiler, its will be the same leaf for each of them, with a different
+    // val)
+    // Get insertion values
+    std::vector<fr> insertion_values;
+    std::vector<fr> insertion_locations;
+    std::array<fr, KERNEL_NEW_NULLIFIERS_LENGTH> new_nullifiers_kernel_1;
+    std::array<fr, KERNEL_NEW_NULLIFIERS_LENGTH> new_nullifiers_kernel_2;
+    for (size_t i = 0; i < NUMBER_OF_NULLIFIERS; ++i) {
+        auto insertion_val = starting_insertion_index + i * spacing;
+
+        if (i < KERNEL_NEW_NULLIFIERS_LENGTH) {
+            new_nullifiers_kernel_1[i] = insertion_val;
+        } else {
+            new_nullifiers_kernel_2[i - KERNEL_NEW_NULLIFIERS_LENGTH] = insertion_val;
+        }
+        insertion_values.push_back(insertion_val);
+        insertion_locations.push_back(NUMBER_OF_NULLIFIERS + i);
+        parallel_insertion_tree.update_element(insertion_val);
+    }
+    // Get the hash paths etc from the insertion values
+    auto witnesses_and_preimages = nullifier_tree.circuit_prep_batch_insert(insertion_values, insertion_locations);
+
+    std::vector<stdlib::merkle_tree::nullifier_leaf> new_nullifier_leaves_preimages = witnesses_and_preimages.first;
+    std::pair<std::vector<std::vector<fr>>, std::vector<uint32_t>> new_nullifier_leaves_witnesses =
+        witnesses_and_preimages.second;
+    auto new_nullifier_leaves_sibling_paths = new_nullifier_leaves_witnesses.first;
+    auto new_nullifier_leave_indexes = new_nullifier_leaves_witnesses.second;
+
+    // Create witness values from this
+    std::array<MembershipWitness<NT, NULLIFIER_TREE_HEIGHT>, NUMBER_OF_NULLIFIERS> new_membership_witnesses;
+    for (size_t i = 0; i < NUMBER_OF_NULLIFIERS; i++) {
+        // create an array of the witness from the depth
+        std::array<fr, NULLIFIER_TREE_HEIGHT> witness_array;
+        std::copy(new_nullifier_leaves_sibling_paths[i].begin(),
+                  new_nullifier_leaves_sibling_paths[i].end(),
+                  witness_array.begin());
+
+        MembershipWitness<NT, NULLIFIER_TREE_HEIGHT> witness = {
+            .leaf_index = uint32_t(new_nullifier_leave_indexes[i]),
+            .sibling_path = witness_array,
+        };
+        new_membership_witnesses[i] = witness;
+
+        // Create circuit compatible preimages - issue created to remove this step
+        NullifierLeafPreimage<NT> preimage = {
+            .leaf_value = new_nullifier_leaves_preimages[i].value,
+            .next_index = new_nullifier_leaves_preimages[i].nextIndex,
+            .next_value = new_nullifier_leaves_preimages[i].nextValue,
+        };
+        new_nullifier_leaves[i] = preimage;
+    }
+
+    // Get expected root with subtrees inserted correctly
+    fr end_root = parallel_insertion_tree.root();
+
+    // Expected end state
+    AppendOnlyTreeSnapshot<NT> nullifier_tree_end_snapshot = {
+        .root = end_root,
+        .next_available_leaf_index = 16,
+    };
+
+    // Get the sibling path, we should be able to use the same path to get to the end root
+    std::vector<fr> sibling_path = parallel_insertion_tree.get_sibling_path(8);
+    std::array<fr, NULLIFIER_SUBTREE_INCLUSION_CHECK_DEPTH> sibling_path_array;
+    // Chop the first 3 levels from the sibling_path
+    sibling_path.erase(sibling_path.begin(), sibling_path.begin() + 3);
+    std::copy(sibling_path.begin(), sibling_path.end(), sibling_path_array.begin());
+
+    // Update our start state
+    // Nullifier trees
+    inputs.start_nullifier_tree_snapshot = nullifier_tree_start_snapshot;
+    inputs.new_nullifiers_subtree_sibling_path = sibling_path_array;
+
+    inputs.kernel_data[0].public_inputs.end.new_nullifiers = new_nullifiers_kernel_1;
+    inputs.kernel_data[1].public_inputs.end.new_nullifiers = new_nullifiers_kernel_2;
+
+    inputs.low_nullifier_leaf_preimages = new_nullifier_leaves;
+    inputs.low_nullifier_membership_witness = new_membership_witnesses;
+
+    return std::make_tuple(inputs, nullifier_tree_start_snapshot, nullifier_tree_end_snapshot);
 }
 
 } // namespace aztec3::circuits::rollup::base::utils
